Strictures for attributes.
Allow attribute-defining classes to declare which attributes they
provide, and consuming classes to declare which providers they use. This
will then set up an automatic map from short names to long names. Detect
conflicts, where two used classes declare the same (short) attribute
name. Create an optional 'strict' mode, where unmapped attributes are
forbidden.

In cases of conflicting short names, if one of the sources was an
explicit mapping specified by the user, that silently wins, and any
conflicting attributes are unavailable in this scope. Otherwise we
report a conflict. The long name in an explicit mapping will not also be
imported under its own name, so conflicts can be resolved by remapping a
different short name to one of the long names.