Support attributes from roles.
Registering method attributes by class is inherently unreliable, because
methods get moved from one class to another (particularly by roles).
Instead, store the attributes against a ref (using a fieldhash so they
go away if the sub does) and then, when we're asked for a list of
methods with attributes, grovel through the symbol table to find all our
methods. As an added advantage, this finds methods by their current
name, rather than the name they were compiled under.

This intentionally only returns attributes for locally-defined methods,
not inherited methods. I could provide an all_method_attrs method which
tracks over mro::get_linear_isa, but then I'd have to uniqify the list.

I've got a nasty hack here to handle method modifiers: CMM keeps a cache
of method with modifiers, and it happens to be 'our', so I'm going
through that to find the original sub. This won't always work,
particularly if a modified method is glob-copied onto another name. It
also won't work if the modifier is modifying an inherited method: in
that case CMM doesn't record the original sub anywhere, so there's
nothing we can do.