A role for resolving TT2ish object paths.
This resolves paths of the form 'foo.bar(arg).baz' by calling the
relevant methods in order. Unlike TT2 only one argument is allowed,
which must be a string; this is because I intend to use this from method
attributes, which have no context in which to resolve variables, and
where it's best to keep things simple. If necessary it's trivial to add
a method to the class to perform more complicated lookups.