Bind Moo attributes to gtk properties.
This adds a 'gtk_prop' key to 'has', which binds the attribute to a
property of a Gtk object (specified by an object path). If this key is
specified an after modifier is applied to the attribute's _trigger method
(which will be created if necessary) which sets the Gtk object's
property. An after modifier is also applied to BUILD, which looks up the
Gtk object, records it for the trigger, and connects to the 'notify'
signal to synchronise in the other direction.