Simplify manual registration of attributes.
Calling ->register_method_attrs means looking up the method with ->can,
parsing the attribute definitions into attributes and arguments, and
resolving the attributes wrt a given set of hints. Wrap all that, first
in a MooX::MethodAttributes method (which requires that the hinthash to
use be passed in), and then in an exported sub. This allows

    method_attrs foo => "Foo(Bar)";

which is perfectly usable.

Currently if our 'foo' method is inherited, the attributes will be
applied to our superclass's method. This is probably Bad, and I should
wrap a null 'around' modifer around it to get a local method to attach
the attribute to.