Deep copy.
The tricky case here is where we need to build a new object, since we
don't know what class it should be. I've handled that by ducking the
issue: we supply a value of {copy_from => $f}, and rely on the copied-to
class to have a coercer that inflates this into an appropriate object.

For one of my use cases, this coercer needs access to $self, which Moo
doesn't support. I've written MooX::Role::Coercer, which allows an
attribute to have a 'coercer' method instead of a 'coerce' sub. It also
provides 'coerce_to' to handle the simple case of 'anything non-blessed
should be passed to <this> constructor'.