A new MakerRole: IgnoreUndef.
This makes Moo treat an undefined attribute as nonexistent, so its
predicate will return false and it will be built if it is fetched. This
is important when working with weakrefs, which self-destruct without
warning.

Also, rename MooX::Role::Coercer to MooX::MakerRole::Coercer, and stop
using CaptainHook to apply a maker role automatically. It was silly
having a 'real' role whose only job was to apply a maker role. Instead,
extend MooX::AccessorMaker so it can apply maker roles to the importing
package.