Disconnect Gtk signals invocant disappears.
Previously I was never disconnecting Gtk signals, meaning that if a
series of objects registered handlers and were then destroyed the
handlers would keep accumulating. Now that Role::WeakClosure can call a
destructor, use that to unregister the handler.

This unregistering will not happen immediately: it won't happen until
the next time the signal handler is called; but that isn't really a
problem. The only way to make it happen immediately would be mucking
about with Variable::Magic, and I'd rather avoid that if I can.