Copiable: include superclasses.
Previously, if a class implemented a Copiable role, subclasses of that
class would not see attributes inherited from that class as copiable.
Fix that with a sprinking of mro::get_linear_isa.

I'm not sure I entirely understand the 'adjustment' logic any more... it
looks suspiciously as though it's changing the registration information
of a 'superrole' based on a 'subrole', which would be Quite Wrong.
Hmmm...