Copiable: remove only/exclude from copy_from.
copy_from used to accept 'only' and 'exclude' options to filter the list
of copied attributes. I think Jacana used to use this, but not any more;
it's a messy idea, so forget it. The whole point here is to get groups
of related attributes into roles to avoid mucking about specifying lists
of them all over the place.